# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_sentnl_inferoute-node_pkg_common.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  pkg_api_orchestrator.Message:
    properties:
      content:
        description: Can be string, array, or map
      role:
        enum:
        - system
        - user
        - assistant
        type: string
    required:
    - content
    - role
    type: object
  pkg_api_orchestrator.OpenAIRequest:
    properties:
      frequency_penalty:
        type: number
      logit_bias:
        additionalProperties:
          format: float64
          type: number
        type: object
      max_tokens:
        type: integer
      messages:
        items:
          $ref: '#/definitions/pkg_api_orchestrator.Message'
        type: array
      model:
        type: string
      "n":
        type: integer
      presence_penalty:
        type: number
      prompt:
        type: string
      sort:
        enum:
        - cost
        - throughput
        type: string
      stop:
        items:
          type: string
        type: array
      stream:
        type: boolean
      temperature:
        type: number
      top_p:
        type: number
      user:
        type: string
    required:
    - model
    type: object
...
host: api.inferoute.com
info:
  contact: {}
  description: API for Inferoute orchestration and providers
  title: Inferoute API
  version: "1.0"
paths:
...
  /v1/chat/completions:
    post:
      consumes:
      - application/json
      description: Handles the complete request flow from consumer to provider
      parameters:
      - description: OpenAI request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/pkg_api_orchestrator.OpenAIRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Provider response
          schema: {}
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/github_com_sentnl_inferoute-node_pkg_common.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_sentnl_inferoute-node_pkg_common.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_sentnl_inferoute-node_pkg_common.ErrorResponse'
      summary: Process an OpenAI-compatible API request
      tags:
      - orchestrator
...
schemes:
- https
- http
swagger: "2.0"